agents/requirements.txt
pydantic-ai[openai,anthropic]>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.0
pyyaml>=6.0
python-dotenv>=1.0
snowflake-connector-python>=3.0
